Polish ToolUsage display components & make sharable
