+++ src/components/ToolUsage/index.ts
@@ -1,4 +1,1 @@
-export * from './CreateToolUsage/CreateToolUsage';
-export * from './DeleteToolUsage/DeleteToolUsage';
-export * from './ManageToolUsage';
-export * from './ToolUsageForm/ToolUsageForm';
+export * from './ToolsUsedList/ToolsUsedList';
