+++ src/scenes/Engagement/LanguageEngagement/LanguageEngagementDetail.graphql
@@ -7,7 +7,7 @@ fragment LanguageEngagementDetail on LanguageEngagement {
   ...EngagementDescription
   ...engagementMilestoneReached
   ...aiAssistedTranslation
-  ...EngagementToolUsage
+  ...toolsUsedList
   ceremony {
     ...CeremonyCard
   }
