--- src/scenes/Engagement/LanguageEngagement/Tools/Tools.graphql
@@ -1,26 +1,0 @@
-fragment EngagementToolUsage on Engagement {
-  ...Id
-  tools {
-    id
-    startDate {
-      canRead
-      canEdit
-      value
-    }
-    tool {
-      ...DisplayTool
-    }
-  }
-}
-
-fragment DisplayTool on Tool {
-  id
-  name {
-    canRead
-    value
-  }
-  aiBased {
-    canRead
-    value
-  }
-}
