--- src/scenes/Engagement/LanguageEngagement/Tools/index.ts
@@ -1,1 +1,0 @@
-export { Tools } from './Tools';
